pyproject.toml
[tool.poetry]
name = "inforce-task"
version = "0.1.0"
description = ""
authors = ["C-Hurtmann <constantine2903@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
djoser = "^2.2.0"
pytest-django = "^4.5.2"
drf-yasg = "^1.21.7"
black = "^23.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
